Stub edns0 payload 1232 for IPv6 and 1432 for IPv4

+++ src/context.h
@@ -118,7 +118,7 @@ struct getdns_context {
 	uint8_t edns_extended_rcode;
 	uint8_t edns_version;
 	uint8_t edns_do_bit;
-	uint16_t edns_maximum_udp_payload_size;
+	int edns_maximum_udp_payload_size; /* -1 is unset */
 
 	getdns_update_callback update_callback;
